requirements.txt
streamlit==1.32.0
pandas==2.2.2
numpy==1.23.5
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
shap==0.41.0
plotly==5.21.0
joblib==1.4.0
google-generativeai==0.4.1


